requirements.txt
backcall==0.1.0
colorama==0.4.1
decorator==4.3.2
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.2
louvain==0.6.1
parso==0.3.4
pickleshare==0.7.5
prompt-toolkit==2.0.9
Pygments==2.3.1
python-igraph==0.7.1.post6
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
